Implemented WriteData, WriteReport, get_last_name, get_first_name, get_pay_rate

+++ src/payroll_data.h
@@ -4,7 +4,7 @@
 
 class PayrollData{
 private:
-    int pay_rate_;
+    double pay_rate_;
     std::string first_name;
     std::string last_name;
